Move EventTrigger generation logic to EventTriggerValueProvider
Consolidate EventTrigger-specific code in one place:
- Add EmitDeclaration static method with all generation logic
- Add FindTargetType helper method
- CreateValuesVisitor now just calls EmitDeclaration
